refactor(deployment): extract semantic policy helpers
