+++ packages/deployment/src/utils/required-access.ts
@@ -1,0 +1,5 @@
+export function missing(required: readonly string[], held: readonly string[] | undefined): boolean {
+  const available = new Set(held ?? []);
+  return required.some(value => !available.has(value));
+}
+
